Some issues fixed
