+++ app/routers/files.py
@@ -4,8 +4,10 @@ from fastapi import APIRouter, File, UploadFile, HTTPException
 
 from starlette import status
 
-from app.models.files import TreeFileTypes, UserGroup
-from app.services.file_service import find_last_file_with_name, increase_last_file_name, generate_tree_json
+from app.models.files import TreeFileTypes
+from app.services.file_service import find_last_file_with_name, increase_last_file_name, generate_tree_json, \
+    get_user_group
+from app.settings import Settings, prem_upload_limit, user_upload_limit
 
 router = APIRouter(prefix="/files", tags=["files"])
 
@@ -23,8 +25,10 @@ async def upload_file(
     file_name = file.filename
     file_size = file.size
     user_group = get_user_group(user_id)
-    if file_size > 50 * 1024 * 1024 and user_group != "premium":
-        raise HTTPException(status_code=413, detail="Upload size is over limit. Max size is 50 MB. Buy Premium.")
+    if file_size > user_upload_limit and user_group != "premium":
+        raise HTTPException(status_code=status.HTTP_413_REQUEST_ENTITY_TOO_LARGE, detail="Upload size is over limit. Max size is 50 MB. Buy Premium.")
+    if file_size > prem_upload_limit:
+        raise HTTPException(status_code=status.HTTP_413_REQUEST_ENTITY_TOO_LARGE, detail="You reached uploading limit. Max size is 100 MB.")
     if file_path in [str(user_id), "/", None]:
         path_to_save = directory_path
     else:
@@ -72,16 +76,3 @@ def get_all_directories_as_dict(user_id: int):
         tree_json = generate_tree_json(base_path, TreeFileTypes.folders)
         return tree_json
     raise HTTPException(status_code=status.HTTP_404_NOT_FOUND, detail="Folders tree not found")
-
-
-users = [
-    {"user_id": 123, "group": UserGroup.PREMIUM},
-    {"user_id": 321, "group": UserGroup.REGULAR},
-]
-
-
-def get_user_group(user_id: int):
-    for user in users:
-        if user["user_id"] == user_id:
-            return user["group"]
-    raise HTTPException(status_code=404, detail="User not found")
