+++ db/database.py
@@ -1,0 +1,6 @@
+from app.models.files import UserGroup
+
+users = [
+    {"user_id": 123, "group": UserGroup.PREMIUM},
+    {"user_id": 321, "group": UserGroup.REGULAR},
+]
